remote: made CombinedCache a composition of Disk and Http Cache
Fixes #8052

The problem: CombinedCache is not working.
The cause: both `OnDiskBlobStore` and `HttpBlobStore` store action results differently from other actions, but not in the same way. The `OnDiskBlobStore` appends "ac_" prefix to the key, while the `HttpBlobStore` stores it in a different directory. Current implementation added "ac_" prefix for both blob stores, which is not correct for the `HttpBlobStore`.
The solution: treat actionResults' `gets` and `puts` in accordance with a particular Blob Store used.

The problem is described in #8052 and the root cause is described in [this](https://github.com/bazelbuild/bazel/issues/8052#issuecomment-483556888) and [this](https://github.com/bazelbuild/bazel/issues/8052#issuecomment-487068204) comments.

Closes #8141.

PiperOrigin-RevId: 247387377